Add condition to suites and tasks
This change introduces the concept of condition, which is an script that
is evaluated before running a test or suite and based on its results the
task is executed or skipped.
The idea of this condition is to avoid adding "if" in the prepare-
execute-restore to exit.
A clear example where it can be used is the test main/refresh-all in
snapd project. In that case the condition should be:

condition: |
    [ "$TRUST_TEST_KEYS" = "true" ]