send keepalive packets every 10 seconds to avoid loose connection
It is common on slow connections to see that the tests fails with a
<kill-timeout reached> error during the prepare phase, specially when
the "quiet" method is used.
The proposed solution is to add a keep alive function to send packets
every 10 seconds and prevent the connection is lost.

Example 1 on localhost
----------------------

+ quiet zypper install -y curl expect git golang-packaging jq lsb-
release netcat-openbsd osc rng-tools xdg-utils xdg-user-dirs

<kill-timeout reached>

Example 2 on travis
-------------------

+ go get -u github.com/kardianos/govendor
+ quiet govendor sync
<kill-timeout reached>