Change the mechanism to detect uptime
This change is moving from "uptime -s" to "date -u --iso-8601=seconds -d
\"`cut -f1 -d. /proc/uptime` seconds ago\"". This is because uptime -s
does not work in opensuse, and in spread it is needed a multiplatform
solution.

As we are not using uptime anymore, it is implemented a delta time to
make sure the sub between prev and curr uptimes is reliable.