client: fix dialOnReboot() if the remote does not reply
The PR#53 broke the timeout handling in dialOnReboot(). The
current code has some problems, especially when the target
does not survive the reboot (e.g. no working network after
the reboot):

- it will try to ssh.Dial() the target and never timeouts
- it will try to get uptime and on error retry forever
- ssh.Dial() is really doing the 5s timeout, on my system its
  more like 60s
- This means that the timeout, relog and retry channel all
  can have data in which case golang picks a random channel.
- This means on timeout in the worst case there is only a 1/3
  chance that the timeout channel is handled which may lead
  to much longer kill-timeouts (because ssh.Dial() takes ~60sec
  for me).

I would love to write a test for this, I think we need to mock
ssh.Dial() for this particular test and ideally getUptime as
well. I will do a followup with this as it will be slightly
more involved.