spread: warn if MATCH exists with an unknown status
We saw some obscure MATCH failures when running spread tests against
snapd. It turns out the spread test is failing because of SIGPIPE
which is triggered by us using "-o pipefail". To make it easier
to find these kinds of issues print a warning in MATCH when an
exit-code != 1 is returned (this is the exit code used when no
match is found by grep).