client: simplify MATCH by using "<<<" (#65)
With SIGPIPE and "echo | grep -q" there is a race condition. If
grep exists before echo is finished writing data then echo will
get a SIGPIPE and this makes MATCH fail (which is very obscure
to the user). To fix this this PR switch to the "<<<" HERE
string (thanks to Chipaca).