spread: run MATCH in a subshell (#67)
We ran into a confusing error when using MATCH in one of our snapd
tests. It runs out the issue can be reproduced with:
```
MATCH () {
    {
        set +ux
    } 2> /dev/null;
    return 0
}
MATCH foo <<< foo
test 1 -eq 2
```
When this is used, the output is:
```
$ bash -eux reproducer.sh
+ MATCH foo
```
I.e. the test line is not displayed.

It turns out the reason is that the MATCH function uses "set +ux"
which turns off the traces globally. When MATCH is used in a
pipe (e.g. "echo foo | MATCH foo") then MATCH is run in a
subshell and the turning off is not criticial. However when
using with here strings for file redirects and no subshell
is used turning off trace globally is an issue. The fix is
simply to run MATCH always in a subshell.