New -workers option used to define the number of workers by system
The number of workers can be set as part of the spread.yaml file and
also by using the `-workers` option, which receives an integer
indicating the number of workers used by all the systems selected for
the run.

In case the `-workers` option is not used, the number of workers used is
the provided for the system in the spread.yaml which is 1 by default.