runner: remove filterDir, default include to "."
The include list was defaulting to all the files and directories in the
project, except for files matching .spread-reuse-*. The list of files is
then forwarded to tar, along with the same exclude rule.

This significantly simplifies the invocation of tar and opens up the
possibility of using .gitignore, which requires tar to read the root of
the directory to work.

Two new spread tests check this behavior.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>