spread: add support to define a custom bios with the qemu backend
For some tests it is important to customize the bios that qemu
uses. One example is the Ubuntu Core 20 project. For the spread
testing there we need a qemu that uses the OVMF uefi bios.

The location of this bios is different on all distributions so
defining it in the spread.yaml seems not appropriate.