Adding maxResults parameters when requesting images on google backend
The ideas of this change is to get up to 500 images from google backend.
This is a workaound for the paginatino issue.

The problem with the order parameter is that because an issue in the
gcloud api, when the orderBy parameter is used, the maxResults does not
work.

As the results are not so many, the idea is to manually order the
results.

I already submitted an issue explaining that bad behaviour:
https://issuetracker.google.com/issues/149183460