spread: add "bios" system option with {legacy,bios,""} options
To boot UC20 in qemu the system needs to be set to boot using
UEFI. However there is currently no way right now to switch a
backend to UEFI or legacy bios mode.

This commit adds the option `bios` to the system and (for now)
allows `legacy`, `uefi`. The empty string or a missing option
means the system default should be used.

This is currently only implemented for qemu.