qemu: make number of CPUs used for qemu configurable
The qemu backend is currently only using a single CPU and there
is no way to configure this. This commit adds support to have a
new `cpu-count` opton in the project that allows customizing this.

The option name is similar to what AWS uses - I'm happy to adjust
as needed. Other options might be `NumCpu` or just `Cpu`.

This currently only used with qemu but could easily be used in
the other backends too if needed.

There is also a slight factory to make it easier testable.