qemu: use disk virtio by default
As discussed in
https://github.com/snapcore/spread/pull/96#pullrequestreview-367801844
the qemu backend should use virtio by default all the time.

This will be useful for UC20 where the UEFI grub bootloader is very
very slow (minutes) when loading the kernel in non-virtio mode.

It's also a good idea in general as it will speedup the tests. The
downside is that projects that hardcode `/dev/sda` will have to do a
one time change to move to `/dev/vda`