Support many external systems in parallel
This change is done to allow the execution on the external systems in
parallel with the objective of speeding up the executions.

It can be done by all including the addresses desired and by defining
the number of workers for the system.

The allocate section does not need any change, just add all the
addresses in the SPREAD_EXTERNAL_ADDRESS var, like:
SPREAD_EXTERNAL_ADDRESS="localhost:8022 localhost:8023"