openstack: use similar image matching strategy as google
The image selection is now as close as possible to what the google
backend is doing. Google has the concept of project specific images
but openstack does not have that. So the lookup is simpler for
openstack: a) check for the exact match first b) check for term
matching (which is a bit more flexible/fuzzy than the previous
substring matching). The extra step that openstack does (but google
does not) is that for multiple term matches the latest image is
selected.