lxd: find the right instance in the list (#181)
Due to prefix or regular expression matching, lxd list may list multiple instances. To account for that, instead of assuming it will always return a slice where the first item is the one we're interested in, iterate over the result and pick the one that matches the name exactly. Return an error if no such match exists.

Co-authored-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>