lxd: Use plan keyword to specify lxd instance type
LXD accepts instance type with the -t argument to `lxc launch`
command, and applies constraints both to containers and virtual
machines.

This is useful in its own right for running test cases with
specific CPU and memory constraints, and also a prerequisite
for consequent patch adding support for launching LXD VMs.

Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>